+++ orix/tests/quaternion/test_rotation.py
@@ -223,11 +223,11 @@ class TestToFromEuler:
         e2 = r.to_euler()
         assert np.allclose(e, e2.data)
 
+    def test_mtex(self, e):
+        _ = Rotation.from_euler(e, convention="mtex")
+
     def test_direction_kwarg(self, e):
         _ = Rotation.from_euler(e, direction="lab2crystal")
-
-    def test_Krakow_Hielscher(self, e):
-        _ = Rotation.from_euler(e, convention="Krakow_Hielscher")
 
     def test_direction_kwarg_dumb(self, e):
         with pytest.raises(ValueError, match="The chosen direction is not one of "):