+++ CHANGELOG.rst
@@ -63,6 +63,7 @@ Removed
 -------
 - `StereographicPlot` methods `azimuth_grid()` and `polar_grid()`.
   Use `stereographic_grid()` instead.
+- `from_euler()` no longer accepts "Krakow_Hielscher" as a convention, use "MTEX" instead.
 
 Fixed
 -----